requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.10.0
Django==1.11.1
django-cors-headers==2.0.2
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
packaging==16.8
pycparser==2.17
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
